requirements.txt
pandas
numpy
scikit-learn
shap
joblib
openpyxl
groq
python-dotenv
